feat: filter, modal, table 관련 함수 module화
