+++ week2/assignment/modules/table.js
@@ -1,0 +1,65 @@
+import members from '../data/members.js';
+
+const tbody = document.querySelector('tbody');
+
+export const loadData = (membersData) => {
+  tbody.innerHTML = '';
+
+  membersData.forEach((member) => {
+    const tr = document.createElement('tr');
+
+    const {
+      name,
+      englishName,
+      github,
+      gender,
+      role,
+      firstWeekGroup,
+      secondWeekGroup,
+    } = member;
+
+    tr.innerHTML = `
+        <td><input type="checkbox" /></td>
+        <td>${name}</td>
+        <td>${englishName}</td>
+        <td><a href="https://github.com/${github}" target="_blank">${github}</a></td>
+        <td>${gender === 'male' ? '남자' : '여자'}</td>
+        <td>${role}</td>
+        <td>${firstWeekGroup}</td>
+        <td>${secondWeekGroup}</td>
+    `;
+
+    tbody.appendChild(tr);
+  });
+
+  updateCheckboxListeners();
+};
+
+const updateCheckboxListeners = () => {
+  const thCheckbox = document.querySelector('#th_checkbox');
+  const checkboxes = document.querySelectorAll('tbody input[type="checkbox"]');
+
+  const isAllChecked = () => {
+    return Array.from(checkboxes).every((checkbox) => checkbox.checked);
+  };
+
+  const handleAllChecked = () => {
+    checkboxes.forEach((checkbox) => {
+      checkbox.checked = thCheckbox.checked;
+    });
+  };
+
+  const updateThCheckbox = () => {
+    thCheckbox.checked = isAllChecked();
+  };
+
+  thCheckbox.addEventListener('click', handleAllChecked);
+
+  checkboxes.forEach((checkbox) => {
+    checkbox.addEventListener('click', updateThCheckbox);
+  });
+};
+
+export const initTable = () => {
+  loadData(members); // 초기 데이터 로드
+};
